requirements.txt
streamlit>=1.32.0
scikit-learn>=1.4.0
pandas>=2.0.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
fastapi>=0.100.0
uvicorn>=0.23.0
groq>=0.4.0
langchain>=0.1.10
langchain-community>=0.0.25
langchain-huggingface>=0.0.1
langchain-groq>=0.0.1
langgraph>=0.0.25
faiss-cpu>=1.8.0
joblib>=1.3.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
